Add SVG exports
Adds a button for downloading the current image as an SVG file.

You may notice that in browsers that have a 'downloaded' files strip
at the bottom of the window (such as Chrome), the lines are redrawn
after the first download. This is just because the inner window
size is changing due to the download strip.